pyproject.toml
[project]
name = "udal-sql-connection-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
]
